Configurable NOTICE file

+++ tools/dotnet_notice_generation.sh
@@ -2,6 +2,22 @@
 
 cd "$(dirname "$0")/.."
 
+
+# Check if the correct number of arguments are provided
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 path_to_notice_file"
+    exit 1
+fi
+
+# Assign arguments to variables
+notice_file_path="$1"
+
+# Check if the notice file exists
+if [ ! -f "$notice_file_path" ]; then
+    echo "Error: JSON file '$notice_file_path' not found"
+    exit 1
+fi
+
 DOTNET_DIRECTORY="cloud_connectors/azure/digital_twins_connector"
 
 dotnet tool install --global dotnet-project-licenses
@@ -9,5 +25,5 @@ mkdir -p "$DOTNET_DIRECTORY/dotnet_licenses_output"
 dotnet-project-licenses -i $DOTNET_DIRECTORY -o -f "$DOTNET_DIRECTORY/dotnet_licenses_output" -u --json -e -c \
 --licenseurl-to-license-mappings "$DOTNET_DIRECTORY/license_url_to_type.json"
 ./tools/dotnet_get_licenses.sh "$DOTNET_DIRECTORY/dotnet_licenses_output/licenses.json" "$DOTNET_DIRECTORY/dotnet_licenses_output"
-./tools/dotnet_append_to_notice.sh NOTICE "$DOTNET_DIRECTORY/dotnet_licenses_output/licenses.json"
+./tools/dotnet_append_to_notice.sh "$notice_file_path" "$DOTNET_DIRECTORY/dotnet_licenses_output/licenses.json"
 rm -rf "$DOTNET_DIRECTORY/dotnet_licenses_output"
